+++ mapitio_auth/codenames/__init__.py
@@ -1,5 +1,6 @@
 from .auditor import auditor
 from .clinic import clinic
+from .export import export
 from .screening import screening
 from .unblinding import unblinding_requestors, unblinding_reviewers
 from .pii import pii