+++ mapitio_subject/admin/biomedical_history_admin.py
@@ -3,6 +3,7 @@ from edc_form_label.form_label_modeladmin_mixin import FormLabelModelAdminMixin
 from edc_model_admin import SimpleHistoryAdmin
 
 from ..admin_site import mapitio_subject_admin
+from ..exim_resources import BiomedicalHistoryResource
 from ..forms import BiomedicalHistoryForm
 from ..models import BiomedicalHistory
 from .fieldsets import (
@@ -21,6 +22,8 @@ class BiomedicalHistoryAdmin(
     )
 
     form = BiomedicalHistoryForm
+
+    resource_class = BiomedicalHistoryResource
 
     fieldsets = biomedical_fieldsets