+++ mapitio_subject/admin/followup_admin.py
@@ -5,15 +5,19 @@ from edc_model_admin import SimpleHistoryAdmin
 from mapitio_subject.admin.fieldsets import comment_fieldset_tuple
 
 from django.utils.safestring import mark_safe
+
 from ..admin_site import mapitio_subject_admin
-from ..forms import FollowUpForm
-from ..models import FollowUp
+from ..exim_resources import FollowupResource
+from ..forms import FollowupForm
+from ..models import Followup
 from .modeladmin import CrfModelAdminMixin
 
 
-@admin.register(FollowUp, site=mapitio_subject_admin)
-class FollowUpAdmin(CrfModelAdminMixin, FormLabelModelAdminMixin, SimpleHistoryAdmin):
-    form = FollowUpForm
+@admin.register(Followup, site=mapitio_subject_admin)
+class FollowupAdmin(CrfModelAdminMixin, FormLabelModelAdminMixin, SimpleHistoryAdmin):
+    form = FollowupForm
+
+    resource_class = FollowupResource
 
     additional_instructions = mark_safe(
         "<span style='color:#ff8000'>Complete for data as of the patient's last attended clinic visit</span>"