+++ mapitio_subject/admin/subject_visit_admin.py
@@ -2,19 +2,25 @@ from django.contrib import admin
 from edc_model_admin import audit_fieldset_tuple, SimpleHistoryAdmin
 from edc_visit_schedule.fieldsets import visit_schedule_fieldset_tuple
 from edc_visit_tracking.modeladmin_mixins import VisitModelAdminMixin
+from import_export.admin import ExportActionMixin
 
 from ..admin_site import mapitio_subject_admin
+from ..exim_resources import SubjectVisitResource
 from ..forms import SubjectVisitForm
 from ..models import SubjectVisit
 from .modeladmin import ModelAdminMixin
 
 
 @admin.register(SubjectVisit, site=mapitio_subject_admin)
-class SubjectVisitAdmin(VisitModelAdminMixin, ModelAdminMixin, SimpleHistoryAdmin):
+class SubjectVisitAdmin(
+    VisitModelAdminMixin, ModelAdminMixin, ExportActionMixin, SimpleHistoryAdmin
+):
 
     show_dashboard_in_list_display_pos = 2
 
     form = SubjectVisitForm
+
+    resource_class = SubjectVisitResource
 
     fieldsets = (
         (