+++ CHANGES.rst
@@ -1,5 +1,12 @@
 Changes
 
+0.1.16
+------
+- change ARV regimen to an autocomplete field from radio
+- add ``Resource`` classes, configure ``admin`` for ``django-import-export``
+- encrypt "confirm" identifier fields on SubjectScreening
+- upgrade to edc==0.1.24
+
 0.1.15
 ------
 - add hiv and indicators followup forms
@@ -8,8 +15,8 @@ Changes
 0.1.14
 ------
 - update arv regimen list data based on "other" values entered
-- change m2ms to blank=True and use form logic `required_if`
-- fix incorrect base class on some admin classes
+- change m2ms to ``blank=True`` and use form logic ``required_if``
+- fix incorrect base class on some ``admin`` classes
 - other minor form logic fixes
 
 0.1.13